[DNM] Serialize Booster before returning in distributed training
There is a race condition between shutting down the Rabit tracker and
serializing the booster. When training single models it seems the
booster is serialized "fast enough", but when training many models in
parallel the race condition is more likely to be triggered.

Work around this by serializing the model to bytes before shutting down
Rabit. While at it dispose the booster to ensure C++ frees associated
memory without having to wait for GC.